style homepage modules a bit
