+++ src/scripts/templates/index.js
@@ -3,5 +3,6 @@
 import '../../styles/templates/index.css'
 
 import {load} from '@shopify/theme-sections';
+import '../sections/product';
 
 load('*');
