+++ src/sections/collection-list.liquid
@@ -1,9 +1,9 @@
 <section class="section helper container mx-auto px-4">
   {% if section.settings.title != blank %}
-    <h2>{{ section.settings.title | escape }}</h2>
+    <h2 class="text-center text-xl">{{ section.settings.title | escape }}</h2>
   {% endif %}
 
-  <div class="grid grid-cols-3 grid-rows-2">
+  <div class="grid gap-10 grid-cols-3 {% if section.blocks.size > 3 %} grid-rows-2 {% else %} grid-rows-1 {% endif %}">
     {% for block in section.blocks limit: section.blocks.size %}
       <div class="" {{ block.shopify_attributes }}>
         {%- assign collection = collections[block.settings.collection] -%}
